Integrated menus and game (read desc.)
I have combined GameShell and main into one file called Temp (this
should probably be renamed) so that working on one would not affect the
other but the integration is available. Feel free to move this into
main; I would prefer that but I don't want to disturb your workspace.